requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.15
python-dotenv>=1.0.0
openai>=1.12.0
requests>=2.31.0
python-dateutil>=2.8.2
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3 